fix compilation issues

+++ src/registry/lib.rs
@@ -70,7 +70,7 @@ pub mod registry {
     )]
     pub struct Agent {
         pub address: AccountId,
-        pub weight: u64,
+        pub weight: u128,
         pub blacklisted: bool,
     }
 
@@ -83,8 +83,8 @@ pub mod registry {
     pub struct AgentUpdated {
         #[ink(topic)]
         agent: AccountId,
-        old_weight: u64,
-        new_weight: u64,
+        old_weight: u128,
+        new_weight: u128,
     }
     #[ink(event)]
     pub struct AgentDeleted {
@@ -107,7 +107,7 @@ pub mod registry {
         // List of all nomination agents including their deployment and relative weight
         pub agents: Vec<Agent>,
         // Sum of nomination agent relative weights
-        pub total_weight: u64,
+        pub total_weight: u128,
         // Permissions for adding agents, updating weights, and removing agents
         pub roles: Mapping<RoleType, Role>,
         // Used for instantiating agents
@@ -233,7 +233,7 @@ pub mod registry {
         pub fn update_agents(
             &mut self,
             agents: Vec<AccountId>,
-            new_updates: Vec<(u64, bool)>,
+            new_updates: Vec<(u128, bool)>,
         ) -> Result<(), RegistryError> {
             let caller = Self::env().caller();
 
@@ -334,7 +334,7 @@ pub mod registry {
         }
 
         #[ink(message)]
-        pub fn get_agents(&self) -> (u64, Vec<Agent>) {
+        pub fn get_agents(&self) -> (u128, Vec<Agent>) {
             (self.total_weight, self.agents.clone())
         }
